+++ src/Controller/InvoiceController.php
@@ -21,6 +21,8 @@ final class InvoiceController extends AbstractController
         InvoiceRepository $invoiceRepository,
         PaginatorInterface $paginator
     ): Response {
+        $this->denyAccessUnlessGranted('ROLE_USER');
+
         $query = $invoiceRepository->createQueryBuilder('i')
             ->orderBy('i.id', 'DESC')
             ->getQuery();
@@ -39,11 +41,19 @@ final class InvoiceController extends AbstractController
     #[Route('/new', name: 'app_invoice_new', methods: ['GET', 'POST'])]
     public function new(Request $request, EntityManagerInterface $entityManager): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_SELLER');
+
         $invoice = new Invoice();
         $form = $this->createForm(InvoiceType::class, $invoice);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
+            foreach ($invoice->getInvoiceItems() as $item) {
+                $price = (float) $item->getPrice();
+                $quantity = (int) $item->getQuantity();
+                $item->setTotal(number_format($price * $quantity, 2, '.', ''));
+            }
+
             $entityManager->persist($invoice);
             $entityManager->flush();
 
@@ -59,6 +69,8 @@ final class InvoiceController extends AbstractController
     #[Route('/{id}', name: 'app_invoice_show', methods: ['GET'])]
     public function show(Invoice $invoice): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_USER');
+
         return $this->render('invoice/show.html.twig', [
             'invoice' => $invoice,
         ]);
@@ -67,10 +79,18 @@ final class InvoiceController extends AbstractController
     #[Route('/{id}/edit', name: 'app_invoice_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, Invoice $invoice, EntityManagerInterface $entityManager): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_SELLER');
+
         $form = $this->createForm(InvoiceType::class, $invoice);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
+            foreach ($invoice->getInvoiceItems() as $item) {
+                $price = (float) $item->getPrice();
+                $quantity = (int) $item->getQuantity();
+                $item->setTotal(number_format($price * $quantity, 2, '.', ''));
+            }
+
             $entityManager->flush();
 
             return $this->redirectToRoute('app_invoice_index', [], Response::HTTP_SEE_OTHER);
@@ -85,7 +105,9 @@ final class InvoiceController extends AbstractController
     #[Route('/{id}', name: 'app_invoice_delete', methods: ['POST'])]
     public function delete(Request $request, Invoice $invoice, EntityManagerInterface $entityManager): Response
     {
-        if ($this->isCsrfTokenValid('delete'.$invoice->getId(), $request->getPayload()->getString('_token'))) {
+        $this->denyAccessUnlessGranted('ROLE_SELLER');
+
+        if ($this->isCsrfTokenValid('delete' . $invoice->getId(), $request->getPayload()->getString('_token'))) {
             $entityManager->remove($invoice);
             $entityManager->flush();
         }
